refactor master detail
